Cambios en el eliminar retiro, mejorado

+++ SistemaFinca/Módulo Ventas/Retiros/VR_Eliminar.cs
@@ -22,50 +22,56 @@ namespace SistemaFinca
         public FormVR_Eliminar()
         {
             InitializeComponent();
+
+
         }
 
-        private void btnRegistrar_Click(object sender, EventArgs e)
+        private void btnEliminar_Click(object sender, EventArgs e)
         {
             if (this.idretiro == "")
             {
                 MessageBox.Show("Seleccione un retiro", "Vuelva a intentar", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 return;
             }
-            using (NpgsqlConnection connection = new NpgsqlConnection(FormLogin.connectionString))
-            {
-                try
-                {
-                    connection.Open();
-                    MessageBox.Show($"{this.idcontrato}, {this.idretiro}", "Vuelva a intentar", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    String commString2 = $"UPDATE contrato SET cantidadretirada = cantidadretirada - (SELECT cantidadleche " +
-                        $"FROM retiro WHERE idretiro = {this.idretiro}) WHERE idcontrato = {this.idcontrato}";
-                    NpgsqlCommand comm2 = new NpgsqlCommand(commString2, connection);
-                    int resultado = comm2.ExecuteNonQuery();
 
-                    String commString = $"DELETE FROM retiro WHERE idretiro = {this.idretiro}";
-                    NpgsqlCommand comm = new NpgsqlCommand(commString, connection);
-                    int resultado2 = comm.ExecuteNonQuery();
-                    if (resultado > 0 && resultado2 > 0)
+            if (MessageBox.Show($"Se eliminará el retiro {this.idretiro} del contrato {this.idcontrato}", "Comprobar", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+            {
+                using (NpgsqlConnection connection = new NpgsqlConnection(FormLogin.connectionString))
+                {
+                    try
                     {
-                        MessageBox.Show("Retiro eliminado exitosamente", "Eliminación exitosa", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        this.idretiro = "";
-                        lstRetiros.Items.Clear();
-                        getRetiros();
-                        return;
+                        connection.Open();
+
+                        String commString2 = $"UPDATE contrato SET cantidadretirada = cantidadretirada - (SELECT cantidadleche " +
+                            $"FROM retiro WHERE idretiro = {this.idretiro}) WHERE idcontrato = {this.idcontrato}";
+                        NpgsqlCommand comm2 = new NpgsqlCommand(commString2, connection);
+                        int resultado = comm2.ExecuteNonQuery();
+
+                        String commString = $"DELETE FROM retiro WHERE idretiro = {this.idretiro}";
+                        NpgsqlCommand comm = new NpgsqlCommand(commString, connection);
+                        int resultado2 = comm.ExecuteNonQuery();
+                        if (resultado > 0 && resultado2 > 0)
+                        {
+                            MessageBox.Show("Retiro eliminado exitosamente", "Eliminación exitosa", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                            this.idretiro = "";
+                            lstRetiros.Rows.Clear();
+                            getRetiros();
+                            return;
+                        }
+                    }
+                    catch (NpgsqlException ex)
+                    {
+                        Console.WriteLine(ex.Message);
+                    }
+                    finally
+                    {
+                        if (connection.State == ConnectionState.Open)
+                        {
+                            connection.Close();
+                        }
                     }
                 }
-                catch (NpgsqlException ex)
-                {
-                    Console.WriteLine(ex.Message);
-                }
-                finally
-                {
-                    if (connection.State == ConnectionState.Open)
-                    {
-                        connection.Close();
-                    }
-                }
-            }
+            }  
         }
 
         private void btnBuscar_Click(object sender, EventArgs e)
@@ -80,6 +86,7 @@ namespace SistemaFinca
 
         private void getRetiros()
         {
+            lstRetiros.Rows.Clear();
             using (NpgsqlConnection connection = new NpgsqlConnection(FormLogin.connectionString))
             {
                 try
@@ -113,6 +120,7 @@ namespace SistemaFinca
                     NpgsqlCommand comm2 = new NpgsqlCommand(commString2, connection);
                     using (NpgsqlDataReader reader = comm2.ExecuteReader())
                     {
+
                         if (!reader.HasRows)
                         {
                             MessageBox.Show("No existen retiros", "Vuelve a intentar", MessageBoxButtons.OK, MessageBoxIcon.Error);
@@ -120,14 +128,15 @@ namespace SistemaFinca
                         }
                         while (reader.Read())
                         {
-                            String cantidad = reader.GetInt32(0).ToString();
-                            String fecha = reader.GetDateTime(1).ToString();
-                            String estado = reader.GetBoolean(2) ? "Pagado" : "No pagado";
-                            ListViewItem item = new ListViewItem(cantidad);
-                            item.SubItems.Add(fecha);
-                            item.SubItems.Add(estado);
-                            item.SubItems.Add(reader.GetInt32(3).ToString());
-                            lstRetiros.Items.Add(item);
+
+                            string cantidad = reader.GetInt32(0).ToString();
+                            string fecha = reader.GetDateTime(1).ToString();
+                            string estado = reader.GetBoolean(2) ? "Pagado" : "No pagado";
+                            string idRetiro = reader.GetInt32(3).ToString();
+
+                            // Agrega una nueva fila al DataGridView con los valores actuales
+                            lstRetiros.Rows.Add(cantidad, fecha, estado, idRetiro);
+                            lstRetiros.ClearSelection();
                         }
                     }
                 }
@@ -145,46 +154,14 @@ namespace SistemaFinca
             }
         }
 
-        public static bool fechaEsValida(string fecha)
+        private void lstRetiros_CellClick(object sender, DataGridViewCellEventArgs e)
         {
-            Regex regex = new Regex(@"^(0[1-9]|[12][0-9]|3[01])/(0[1-9]|1[0-2])/\d{2}$");
-            return regex.IsMatch(fecha);
-        }
-
-        public static bool cantidadEsValida(string cantidad)
-        {
-            Regex regex = new Regex(@"^[1-9]\d{0,3}$");
-            return regex.IsMatch(cantidad);
-        }
-
-        public static bool fechaEsPosterior(string fechaInicioStr, string fechaFinStr)
-        {
-            DateTime fechaInicio, fechaFin;
-
-            if (DateTime.TryParseExact(fechaInicioStr, "dd/MM/yy", null, System.Globalization.DateTimeStyles.None, out fechaInicio) &&
-                DateTime.TryParseExact(fechaFinStr, "dd/MM/yy", null, System.Globalization.DateTimeStyles.None, out fechaFin))
+            if (e.RowIndex > -1)
             {
-                return fechaFin > fechaInicio;
-            }
-            return false;
-        }
+                lstRetiros.ClearSelection();
+                lstRetiros.Rows[e.RowIndex].Selected = true;
 
-        public static string ConvertirFecha(string fechaOriginalStr)
-        {
-            DateTime fechaOriginal;
-
-            if (DateTime.TryParseExact(fechaOriginalStr, "dd/MM/yy", null, System.Globalization.DateTimeStyles.None, out fechaOriginal))
-            {
-                return fechaOriginal.ToString("yyyy-MM-dd");
-            }
-            return fechaOriginalStr;
-        }
-
-        private void lstRetiros_ItemSelectionChanged(object sender, ListViewItemSelectionChangedEventArgs e)
-        {
-            if (lstRetiros.SelectedItems.Count > 0)
-            {
-                this.idretiro = lstRetiros.SelectedItems[0].SubItems[3].Text;
+                this.idretiro = lstRetiros.Rows[e.RowIndex].Cells[3].Value.ToString();             
             }
         }
     }
